Adding e2e test cases for personal sign page
