+++ app/components/Views/confirmations/components/Confirm/Info/Shared/InfoRowOrigin/InfoRowOrigin.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 
+import { ConfirmationPageSectionsSelectorIDs } from '../../../../../../../../../e2e/selectors/Confirmation/ConfirmationView.selectors';
 import { strings } from '../../../../../../../../../locales/i18n';
 import useApprovalRequest from '../../../../../hooks/useApprovalRequest';
 import InfoSection from '../../../../UI/InfoRow/InfoSection';
 import InfoRow from '../../../../UI/InfoRow';
 import DisplayURL from '../../../../UI/InfoRow/InfoValue/DisplayURL';
-import { ConfirmationPageSectionsSelectorIDs } from '../../../../../../../../../e2e/selectors/Confirmation/ConfirmationView.selectors';
 
 const InfoRowOrigin = () => {
   const { approvalRequest } = useApprovalRequest();
