+++ e2e/specs/confirmations/signatures/ethereum-sign.spec.js
