+++ e2e/specs/confirmations/signatures/security-alert-signatures.mock.spec.js
