+++ e2e/specs/confirmations/signatures/typed-sign-v4.spec.js
