+++ e2e/specs/confirmations/signatures/typed-sign.spec.js
